Migrated internals like boxes/arrows/items to a single Model object for simplicity.
Thus, the Model is used for saving and loading.

A State object contains a Model as well as current box (prompt) and the
available choices.

Changes to the Model fire events handled by registered DirtyListener
objects (the Editor and the Canvas).

In the future, these DirtyListener events may have more information.