Moved files into typical Eclipse configuration.
Source files are all in src and class files remain in bin.

Added project files into a projects directory and a save file into a
saves directory.

These are now easier to access from the CLI game.

Fixed some bugs and formatting in the CLI game.

Also, added support for newlines and long text.  New lines are displayed
in tooltips and in radio button text through the use of Swing's limited
HTML support.